AMP-30620 AMP-30606-Issue- tools> ws info. manager cannotremove member of ws(Cast issue) AMP- 30658 -Issue- activity file upload functionality in related document is broken(Database connection timeout) AMP-30656-Issue- cannot add computation ws based on filter-(CSRF) AMP-30657 Issue- activity statuses and other dropdown fields are broken(Database sequence issue. Not yet fixed)

+++ amp/WEB-INF/src/org/digijava/kernel/security/ResourcePermission.java
@@ -94,7 +94,7 @@ public abstract class ResourcePermission
             if (v == null) {
                 throw new RuntimeException("Unkonow security action: " + action);
             } else {
-                mask = mask | modifySecurityAction(( (Integer) v).intValue());
+                mask = mask | modifySecurityAction(v);
             }
         }
         return mask;
@@ -196,7 +196,7 @@ public abstract class ResourcePermission
         Iterator iter = getCodesToActions().entrySet().iterator();
         while (iter.hasNext()) {
             Map.Entry item = (Map.Entry)iter.next();
-            int code = ((Integer)item.getKey()).intValue();
+            int code = (Integer) item.getKey();
             String name = (String)item.getValue();
 
             if ((code | tmpMask) == tmpMask) {
